test trigger

+++ GarageCreditCeleste/Acheter.cs
@@ -77,7 +77,7 @@ namespace GarageCreditCeleste
                             // Parcourir les résultats
                             while (oReader.Read())
                             {
-                                Voiture uneVoiture = new Voiture(Convert.ToString(oReader[1]), Convert.ToString(oReader[2]), Convert.ToInt16(oReader[3]), Convert.ToInt16(oReader[5]), Convert.ToString(oReader[6]), Convert.ToInt16(oReader[7]), Convert.ToString(oReader[9]), Convert.ToBoolean(oReader[8]), Convert.ToInt16(oReader[4]));
+                                Voiture uneVoiture = new Voiture(Convert.ToString(oReader[1]), Convert.ToString(oReader[2]), Convert.ToInt16(oReader[3]), Convert.ToInt16(oReader[5]), Convert.ToString(oReader[6]), Convert.ToInt16(oReader[7]), Convert.ToString(oReader[9]), Convert.ToBoolean(oReader[8]), Convert.ToDouble(oReader[4]));
                                 lesVoitures.Add(uneVoiture);
                             }
                         }
